move log helper to lib
