+++ rest/helpers.go
@@ -2,12 +2,7 @@ package rest
 
 import (
 	"fmt"
-	"io"
-	"io/ioutil"
 	"net/http"
-	"os"
-
-	"github.com/flavioribeiro/gonfig"
 )
 
 // HTTPError is a helper to return errors on handlers
@@ -22,26 +17,4 @@ func JSONHandler(actual http.Handler) http.Handler {
 		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 		actual.ServeHTTP(w, r)
 	})
-}
-
-// GetLogOutput returns the output we want to use
-// for http requests log
-func GetLogOutput() io.Writer {
-	var logOutput io.Writer
-	currentDir, _ := os.Getwd()
-	cfg, _ := gonfig.FromJsonFile(currentDir + "/config.json")
-	logfile, _ := cfg.GetString("LOGFILE", "")
-	if logfile == "" {
-		logOutput = ioutil.Discard
-	} else {
-		fmt.Println("Logging requests on", logfile)
-		f, err := os.Create(logfile)
-		if err != nil {
-			panic(err)
-		}
-
-		logOutput = f
-	}
-
-	return logOutput
 }
